Fallback chain if no connection
There is still the issue of being unable to connect to a new network if the network has not been already created in the wallet. This should be able to be resolved with a raw wallet_addEthereumChain RPC call, however, we'll need to have a catch alongside the wagmi switchNetwork that only runs when this error is returned.