ABIs in loader and hardhat switch disabled
Importing ABIs and contract info is placed within the index of each route behind a loader to handle server side.

This ABI and contract data is passed to a useContext wrapper that can then be accessed within any component inside the route tree. (Open to feedback on if we actually need the context wrapper but it made for cleaner code in my head).

Finally, I have tried to solve this switch network issue that is specific to hardhat for much too long. As of now, the button is hidden if our operating network is hardhat. I think it is worthy of its own ticket depending on the feedback I get from my metamask and wagmi github issues.