feat: update to wallet connect v2
This upgrades RainbowKit and wagmi to the last version that supported Ethers to avoid the need for refactoring everything.

Right now, generation of contract interactions was based around `typechain` that uses `ethers-v5` meaning migration is not exactly straightforward given `wagmi` replaced `ethers` with `viem`

Given a fix for interaction generation migration can be done with ease.